Time: 25 ms (60.65%), Space: 19.3 MB (46.45%) - LeetHub

+++ 0283-move-zeroes/0283-move-zeroes.cpp
@@ -2,18 +2,8 @@ class Solution {
 public:
     void moveZeroes(vector<int>& nums) {
         
-//        auto j = nums.size();
-//        for (auto idx = nums.size(); idx-- != 0;)
-//        {
-//            if (nums[idx] != 0)
-//            {
-//                --j;
-//                std::swap(nums[idx], nums[j]);
-//            }
-//        }
-        
-        auto j = -1;
-        for (auto idx = 0; idx != nums.size(); ++idx)
+        auto j = -1;        // pointer for zero element
+        for (auto idx = 0; idx != static_cast<int>(nums.size()); ++idx)
         {
             if (nums[idx] != 0)
             {
